refactor: organize validators into scrapers/, scoring/, clients/ subdirectories
Move 12 top-level files into logical subdirectories to reduce clutter.
Inline STEEPNESS/FACTOR constants into performance_reward.py and delete constants.py.
