--- neurons/validators/constants.py
@@ -1,2 +1,0 @@
-STEEPNESS = 0.1
-FACTOR = 2  # a scaling factor
