+++ neurons/validators/proxy/uid_manager.py
@@ -3,7 +3,7 @@ from typing import List
 
 from bittensor.core.metagraph import AsyncMetagraph
 
-from neurons.validators.weights import EMISSION_CONTROL_HOTKEY
+from neurons.validators.scoring.weights import EMISSION_CONTROL_HOTKEY
 
 
 class UIDManager:
